Two comments
I have two smaller comments and thought it would be the right thing to add them here as pull request, which is of course not intended to be merged.

Apart from the two comments I think your timeline is surprisingly (and challengingly) concrete ...

But there's not much that needs to be changed, I think. Referring to the GoogleGroup discussion is appropriate because that's public and quite detailed.